MAINT: Move transforms module to core sub-package
