+++ doc/source/dev_reference/core.rst
@@ -6,3 +6,4 @@ Core classes and functions.
 
 .. automodule:: pyart.core.grid
 .. automodule:: pyart.core.radar
+.. automodule:: pyart.core.transforms
