+++ doc/source/dev_reference/util.rst
@@ -6,5 +6,4 @@ Miscellaneous utility functions.
 
 .. automodule:: pyart.util.circular_stats
 .. automodule:: pyart.util.hildebrand_sekhon
-.. automodule:: pyart.util.transforms
 .. automodule:: pyart.util.xsect
