+++ pyart/map/grid_mapper.py
@@ -28,8 +28,8 @@ import numpy as np
 import scipy.spatial
 
 from ..config import get_fillvalue
-from ..util.transforms import corner_to_point
-from ..util.transforms import antenna_to_cartesian
+from ..core.transforms import corner_to_point
+from ..core.transforms import antenna_to_cartesian
 from ..core.grid import Grid
 from ..core.radar import Radar
 from ..filters import GateFilter, moment_based_gate_filter
